requirements-test.txt
# Test dependencies
-r requirements-ai.txt

# Testing
pytest==7.4.0
pytest-django==4.5.2
pytest-cov==4.1.0
pytest-mock==3.11.1
pytest-asyncio==0.21.1
pytest-xdist==3.3.1
coverage==7.2.7
factory-boy==3.3.0
model-bakery==1.10.1

# Test database
pytest-postgresql==4.1.0
pytest-redis==2.2.0

# Code quality
black==23.7.0
isort==5.12.0
flake8==6.1.0
mypy==1.5.0
bandit==1.7.5
safety==2.3.5

# Documentation
sphinx==7.0.1
sphinx-rtd-theme==1.2.2
sphinx-autodoc-typehints==1.23.0

# Development tools
debugpy==1.7.0
ipython==8.14.0
jedi==0.19.0

# Mocking
responses==0.23.3
freezegun==1.2.2

# API testing
pytest-factoryboy==2.5.1
pytest-bdd==7.0.1
pytest-html==4.1.1

# Performance testing
pytest-benchmark==4.0.0

# Debugging
pdbpp==0.10.3
ipdb==0.13.13

# Test reporting
pytest-html==4.1.1
pytest-sugar==0.9.7

# Test data generation
faker==19.3.0

# Django test tools
django-test-plus==2.0.1
django-webtest==1.9.10

# API documentation
drf-yasg==1.21.5
drf-spectacular==0.26.2

# Security testing
bandit==1.7.5
safety==2.3.5

# Browser automation
selenium==4.11.2
pytest-selenium==4.1.0

# For testing file uploads
Pillow==10.0.0

# For testing async code
pytest-asyncio==0.21.1

# For testing Django channels
pytest-django-channels==0.1.3

# For testing REST APIs
drf-test-utils==1.4.0
pytest-django-queries==1.2.0

# For testing performance
pytest-benchmark==4.0.0

# For testing security
dj-rest-auth==5.0.10
drf-passwordless==1.6.3

# For testing OAuth
python-jose==3.3.0
python-jose[cryptography]==3.3.0

# For testing JWT
djangorestframework-simplejwt==5.2.2

# For testing CORS
django-cors-headers==4.2.0

# For testing Redis
django-redis==5.2.0

# For testing Celery
pytest-celery==0.0.0

# For testing file storage
minio==7.1.16

# For testing PDF generation
reportlab==4.0.4

# For testing Excel files
openpyxl==3.1.2

# For testing CSV files
pandas==2.0.3

# For testing JSON
pytest-json-report==1.5.0

# For testing YAML
pyyaml==6.0.1

# For testing XML
defusedxml==0.7.1

# For testing images
Pillow==10.0.0

# For testing audio
pydub==0.25.1

# For testing video
opencv-python-headless==4.8.0.74

# For testing machine learning models
tensorflow==2.13.0
torch==2.0.1
transformers==4.31.0

# For testing NLP
spacy==3.6.0
nltk==3.8.1

# For testing computer vision
opencv-python-headless==4.8.0.74
mediapipe==0.10.0

# For testing async code
pytest-asyncio==0.21.1

# For testing WebSockets
channels==4.0.0
channels-redis==4.1.0

# For testing GraphQL
graphene-django==3.1.0
graphql-core==3.2.3

# For testing gRPC
grpcio==1.56.0
grpcio-tools==1.56.0

# For testing WebSockets
websockets==11.0.3

# For testing async database access
aiopg==1.4.0
aioredis==2.0.1

# For testing async HTTP clients
httpx==0.24.1

# For testing async task queues
celery==5.3.1
redis==4.5.5

# For testing async cache
aiohttp==3.8.5

# For testing async file I/O
aiofiles==23.1.0

# For testing async database migrations
alembic==1.11.3

# For testing async ORM
tortoise-orm==0.19.3

# For testing async web frameworks
fastapi==0.99.1
uvicorn==0.23.1

# For testing async message brokers
kafka-python==2.0.2

# For testing async search
elasticsearch[async]==8.8.1

# For testing async logging
python-json-logger==2.0.7

# For testing async monitoring
prometheus-client==0.17.0

# For testing async tracing
opentelemetry-api==1.17.0
opentelemetry-sdk==1.17.0

# For testing async metrics
aioinflux==0.10.0

# For testing async job queues
arq==0.26.0

# For testing async web scraping
playwright==1.35.0

# For testing async database drivers
asyncpg==0.27.0

# For testing async message queues
aiormq==6.7.1

# For testing async cache drivers
aiocache==0.12.0

# For testing async file storage
aiofile==0.8.0

# For testing async email
aioimaplib==1.0.1

# For testing async search engines
aiosearch==0.2.0

# For testing async logging handlers
aiofiles==23.1.0

# For testing async task scheduling
apscheduler==3.10.1

# For testing async web servers
hypercorn==0.14.4

# For testing async database migrations
alembic==1.11.3

# For testing async ORM
tortoise-orm==0.19.3

# For testing async web frameworks
fastapi==0.99.1
uvicorn==0.23.1

# For testing async message brokers
kafka-python==2.0.2

# For testing async search
elasticsearch[async]==8.8.1

# For testing async logging
python-json-logger==2.0.7

# For testing async monitoring
prometheus-client==0.17.0

# For testing async tracing
opentelemetry-api==1.17.0
opentelemetry-sdk==1.17.0

# For testing async metrics
aioinflux==0.10.0

# For testing async job queues
arq==0.26.0

# For testing async web scraping
playwright==1.35.0

# For testing async database drivers
asyncpg==0.27.0

# For testing async message queues
aiormq==6.7.1

# For testing async cache drivers
aiocache==0.12.0

# For testing async file storage
aiofile==0.8.0

# For testing async email
aioimaplib==1.0.1

# For testing async search engines
aiosearch==0.2.0

# For testing async logging handlers
aiofiles==23.1.0

# For testing async task scheduling
apscheduler==3.10.1

# For testing async web servers
hypercorn==0.14.4
